refactor: 숫자들 상수화 처리
