+++ apps/client/src/pages/time-table/components/time-table-section/time-table-section.css.ts
