+++ apps/client/src/pages/time-table/constants/index.ts
@@ -1,0 +1,3 @@
+export const HALF_HOUR_TO_MINUTES = 30;
+export const TIME_SLOT_HEIGHT_5_MIN = 5;
+export const ONE_HOUR_TO_MINUTES = 60;
