+++ apps/client/src/pages/time-table/utils/index.ts
@@ -1,3 +1,8 @@
+import {
+  TIME_SLOT_HEIGHT_5_MIN,
+  ONE_HOUR_TO_MINUTES,
+} from '@pages/time-table/constants';
+
 export const generateTableRow = (startTime: number) => {
   return Array.from({ length: 24 - startTime }, (_, idx) => startTime + idx);
 };
@@ -7,8 +12,8 @@ export const parseTimeString = (timeString: string): number[] => {
 };
 
 export const calcPosition = (totalMin: number, minutesFromOpen: number) => {
-  const top = (minutesFromOpen / 5) * 0.74;
-  const diff = (totalMin / 5) * 0.74;
+  const top = (minutesFromOpen / 5) * TIME_SLOT_HEIGHT_5_MIN;
+  const diff = (totalMin / 5) * TIME_SLOT_HEIGHT_5_MIN;
   return { top, diff };
 };
 
@@ -18,7 +23,7 @@ export const calcTotalMinutes = (
   endHour: number,
   endMin: number,
 ) => {
-  return endHour * 60 + endMin - (startHour * 60 + startMin);
+  return endHour * 60 + endMin - (startHour * ONE_HOUR_TO_MINUTES + startMin);
 };
 
 export const calcMinutesFromOpen = (
@@ -27,7 +32,7 @@ export const calcMinutesFromOpen = (
   openHour: number,
   openMin: number,
 ) => {
-  const startTotalMin = startHour * 60 + startMin;
-  const openTotalMin = openHour * 60 + openMin;
+  const startTotalMin = startHour * ONE_HOUR_TO_MINUTES + startMin;
+  const openTotalMin = openHour * ONE_HOUR_TO_MINUTES + openMin;
   return startTotalMin - openTotalMin;
 };
